refactor/pong-together/frontend#59: 모달 코드 리팩토링
